requirements2.txt
Package                Version
---------------------- ----------------
bcc                    0.18.0
bcrypt                 4.0.1
blinker                1.4
Brlapi                 0.8.3
certifi                2020.6.20
cffi                   1.16.0
chardet                4.0.0
chrome-gnome-shell     0.0.0
click                  8.0.3
colorama               0.4.4
command-not-found      0.3
cryptography           3.4.8
cupshelpers            1.0
dbus-python            1.2.18
defer                  1.0.6
distro                 1.7.0
future                 0.18.3
greenlet               1.1.2
hidpidaemon            18.4.6
httplib2               0.20.2
idna                   3.3
importlib-metadata     4.6.4
jeepney                0.7.1
Jinja2                 3.1.2
junos-eznc             2.6.8
kernelstub             3.1.4
keyring                23.5.0
language-selector      0.1
launchpadlib           1.10.16
lazr.restfulclient     0.14.4
lazr.uri               1.0.6
louis                  3.20.0
lxml                   4.9.3
macaroonbakery         1.3.1
MarkupSafe             2.1.3
more-itertools         8.10.0
msgpack                1.0.3
mypy-extensions        1.0.0
napalm                 4.1.0
ncclient               0.6.13
netaddr                0.8.0
netifaces              0.11.0
netmiko                4.2.0
netutils               1.6.0
nornir                 3.4.1
nornir-napalm          0.4.0
nornir-netbox          0.3.0
nornir-utils           0.2.0
ntc-templates          3.5.0
oauthlib               3.2.0
paramiko               3.3.1
pip                    22.0.2
pop-transition         1.1.2
protobuf               3.12.4
pycairo                1.20.1
pycparser              2.21
pycups                 2.0.1
pydbus                 0.6.0
pyeapi                 1.0.2
PyGObject              3.42.1
PyJWT                  2.3.0
pymacaroons            0.13.0
PyNaCl                 1.5.0
pynvim                 0.4.2
pyparsing              2.4.7
pyRFC3339              1.1
pyserial               3.5
python-apt             2.4.0+ubuntu2
python-debian          0.1.43+ubuntu1.1
python-dotenv          1.0.0
python-gnupg           0.4.8
python-xlib            0.29
pytz                   2022.1
pyxdg                  0.27
PyYAML                 5.4.1
repolib                2.2.1
repoman                1.4.0
requests               2.25.1
ruamel.yaml            0.17.35
ruamel.yaml.clib       0.2.8
scp                    0.14.5
SecretStorage          3.3.1
sessioninstaller       0.0.0
setuptools             59.6.0
six                    1.16.0
systemd-python         234
textfsm                1.1.3
transitions            0.9.0
ttp                    0.9.5
ttp-templates          0.3.5
typing_extensions      4.8.0
ubuntu-advantage-tools 8001
ubuntu-drivers-common  0.0.0
ufw                    0.36.1
urllib3                1.26.5
wadllib                1.3.6
wheel                  0.37.1
xdg                    5
xkit                   0.0.0
yamlordereddictloader  0.4.2
zipp                   1.0.0
